Split middleware into its own files
